+++ client/src/launcher.js
@@ -105,7 +105,6 @@ window.addEventListener("error", function(e) {
     "src/utils/message-character.js",
     "src/input/keyboard.js",
     "src/input/mouse.js",
-    "src/input/touch.js",
     "src/entities/slot.js",
     "src/entities/container.js",
     "src/entities/equipment.js",